Use APIClient & direct download for projects:backup
